Separate stats reporters for storage and other serviced metrics
